+++ src/components/NavbarMobile.jsx
@@ -57,7 +57,7 @@ const NavbarMobile = ({ isLogged, setIsLogged, updateIsLogged, path, setPath, us
                 </ul>
                 <div className="NavbarMobile--Bottom">
                     <Link to="/credits" style={ path === "/credits" ? selectedStyle : {} }>Credits</Link>
-                    { isLogged && <Link to="/" onClick={handleLogout} style={{ marginBottom: "20px" }}>Logout</Link> }
+                    { isLogged && <Link to="/" onClick={() => handleLogout(updateIsLogged)} style={{ marginBottom: "20px" }}>Logout</Link> }
                     { isLogged && <p>Logged in as {username}</p> }
                     <p>© 2022 Bangerify. All rights reserved.</p>
                 </div>