+++ src/CHANGELOG.md
@@ -1,3 +1,7 @@
 # 26-03-2023
 - markdown links are now opening in new tab
 - @{nick} in post is now link to the profile
+
+# 26-03-2023.2
+- markdown and profile links (@) in comments
+- can login using email